SAS code for a multi site study
on comparisons between patients diagnosed with PTSD in primary versus mental health care published in 'Journal of Clinical Psychology in Medical Setting (JOCS)' which can be found in PubMed: https://www.ncbi.nlm.nih.gov/pubmed/32048114

This is the first part of the code to create the final data set. This code was run by 5 medical sites and then the data was aggregated for analysis.